+++ packages/cmd/src/CommandProcessor.ts
@@ -122,6 +122,13 @@ export class CommandProcessor {
     private mLogger: Logger = Logger.getImperativeLogger();
 
     /**
+     * Current working directory
+     * @private
+     * @memberof CommandProcessor
+     */
+    private mCwd: string;
+
+    /**
      * Creates an instance of CommandProcessor.
      * @param {ICommandProcessorParms} params - See the interface for details.
      * @memberof CommandProcessor
@@ -147,6 +154,7 @@ export class CommandProcessor {
         ImperativeExpect.keysToBeDefinedAndNonBlank(params, ["promptPhrase"], `${CommandProcessor.ERROR_TAG} No prompt phrase supplied.`);
         ImperativeExpect.keysToBeDefinedAndNonBlank(params, ["rootCommandName"], `${CommandProcessor.ERROR_TAG} No root command supplied.`);
         ImperativeExpect.keysToBeDefinedAndNonBlank(params, ["envVariablePrefix"], `${CommandProcessor.ERROR_TAG} No ENV variable prefix supplied.`);
+        this.mCwd = process.cwd();
         // TODO - check if the command definition passed actually exists within the full command definition tree passed
     }
 
@@ -170,10 +178,6 @@ export class CommandProcessor {
         return this.mCommandLine;
     }
 
-    // set commandLine(command: string) {
-    //     this.mCommandLine = command;
-    // }
-
     /**
      * Accessor for the environment variable prefix
      * @readonly
@@ -193,7 +197,6 @@ export class CommandProcessor {
     get promptPhrase(): string {
         return this.mPromptPhrase;
     }
-
 
     /**
      * Accessor for the help generator passed to this instance of the command processor
@@ -349,6 +352,13 @@ export class CommandProcessor {
 
         // Build the response object, base args object, and the entire array of options for this command
         // Assume that the command succeed, it will be marked otherwise under the appropriate failure conditions
+        if (params.arguments.dcd) {
+
+            // NOTE(Kelosky): we adjust `cwd` and do not restore it, so that multiple simultaneous requests from the same
+            // directory will operate without unexpected chdir taking place.  Multiple simultaneous requests from different
+            // directories may cause unpredictable results
+            process.chdir(params.arguments.dcd as string)
+        }
         const prepareResponse = this.constructResponseObject(params);
         prepareResponse.succeeded();